chore: uncluttered decisiontree.py
