+++ random_tree_models/decisiontree.py
@@ -21,6 +21,10 @@ from rich.tree import Tree
 from sklearn.utils.multiclass import check_classification_targets, unique_labels
 from sklearn.utils.validation import check_array, check_is_fitted, check_X_y
 
+import random_tree_models.leafweights as leafweights
+import random_tree_models.scoring as scoring
+import random_tree_models.utils as utils
+
 
 @dataclass(validate_on_init=True)
 class SplitScore:
@@ -57,13 +61,6 @@ class Node:
         return self.left is None and self.right is None
 
 
-@dataclass
-class TreeGrowthParameters:
-    max_depth: StrictInt = 42
-    min_improvement: StrictFloat = 0.0
-    lam: StrictFloat = 0.0
-
-
 def check_is_baselevel(
     y: np.ndarray, node: Node, depth: int, max_depth: int = None
 ) -> T.Tuple[bool, str]:
@@ -82,191 +79,6 @@ def check_is_baselevel(
         return (False, "")
 
 
-def check_y_and_target_groups(y: np.ndarray, target_groups: np.ndarray = None):
-    n = len(y)
-    if n == 0:
-        raise ValueError(f"{n=}, expected at least one target value")
-
-    if target_groups is not None and len(target_groups) != n:
-        raise ValueError(f"{y.shape=} != {target_groups.shape=}")
-
-
-def calc_variance(y: np.ndarray, target_groups: np.ndarray, **kwargs) -> float:
-    """Calculates the variance of a split"""
-
-    check_y_and_target_groups(y, target_groups=target_groups)
-
-    n = len(y)
-
-    if len(np.unique(target_groups)) == 1:
-        return -np.var(y)
-
-    w_left = target_groups.sum() / n
-    w_right = 1.0 - w_left
-
-    var_left = np.var(y[target_groups])
-    var_right = np.var(y[~target_groups])
-
-    var = w_left * var_left + w_right * var_right
-    return -var
-
-
-def entropy(y: np.ndarray) -> float:
-    "Calculates the entropy across target values"
-
-    n = len(y)
-    check_y_and_target_groups(y)
-
-    unique_ys = np.unique(y)
-    ns = np.array([(y == y_val).sum() for y_val in unique_ys], dtype=int)
-
-    ps = ns / float(n)
-
-    if (ps == 1).any():
-        return 0
-
-    mask_ne0 = ~np.isclose(ps, 0)
-
-    h = ps[mask_ne0] * np.log2(ps[mask_ne0])
-    h = h.sum()
-
-    return h
-
-
-def calc_entropy(y: np.ndarray, target_groups: np.ndarray, **kwargs) -> float:
-    """Calculates the entropy of a split"""
-
-    check_y_and_target_groups(y, target_groups=target_groups)
-
-    w_left = target_groups.sum() / len(target_groups)
-    w_right = 1.0 - w_left
-
-    h_left = entropy(y[target_groups]) if w_left > 0 else 0
-    h_right = entropy(y[~target_groups]) if w_right > 0 else 0
-
-    h = w_left * h_left + w_right * h_right
-    return h
-
-
-def gini_impurity(y: np.ndarray) -> float:
-    "Calculates the gini impurity across target values"
-
-    check_y_and_target_groups(y)
-
-    n = len(y)
-
-    unique_ys = np.unique(y)
-    ns = np.array([(y == y_val).sum() for y_val in unique_ys], dtype=int)
-
-    ps = ns / float(n)
-
-    if (ps == 1).any():
-        return 0
-
-    mask_ne0 = ~np.isclose(ps, 0)
-
-    g = ps[mask_ne0] * (1 - ps[mask_ne0])
-    g = g.sum()
-
-    return -g
-
-
-def calc_gini_impurity(
-    y: np.ndarray, target_groups: np.ndarray, **kwargs
-) -> float:
-    """Calculates the gini impurity of a split
-
-    Based on: https://scikit-learn.org/stable/modules/tree.html#classification-criteria
-    """
-
-    check_y_and_target_groups(y, target_groups=target_groups)
-
-    w_left = target_groups.sum() / len(target_groups)
-    w_right = 1.0 - w_left
-
-    g_left = gini_impurity(y[target_groups]) if w_left > 0 else 0
-    g_right = gini_impurity(y[~target_groups]) if w_right > 0 else 0
-
-    g = w_left * g_left + w_right * g_right
-    return g
-
-
-def xgboost_split_score(
-    g: np.ndarray, h: np.ndarray, growth_params: TreeGrowthParameters
-) -> float:
-    "Equation 7 in Chen et al 2016, XGBoost: A Scalable Tree Boosting System"
-    check_y_and_target_groups(g, target_groups=h)
-
-    top = g.sum()
-    top = top * top
-
-    bottom = h.sum() + growth_params.lam
-
-    if bottom == 0:
-        return 0.0
-
-    score = top / bottom
-    return -score
-
-
-def calc_xgboost_split_score(
-    y: np.ndarray,
-    target_groups: np.ndarray,
-    g: np.ndarray,
-    h: np.ndarray,
-    growth_params: TreeGrowthParameters,
-    **kwargs,
-) -> float:
-    """Calculates the xgboost general version score of a split with loss specifics in g and h.
-
-    Based on: https://scikit-learn.org/stable/modules/tree.html#classification-criteria
-    """
-
-    check_y_and_target_groups(g, target_groups=target_groups)
-    check_y_and_target_groups(h, target_groups=target_groups)
-
-    n_left = target_groups.sum()
-    n_right = len(target_groups) - n_left
-
-    score_left = (
-        xgboost_split_score(g[target_groups], h[target_groups], growth_params)
-        if n_left > 0
-        else 0
-    )
-    score_right = (
-        xgboost_split_score(g[~target_groups], h[~target_groups], growth_params)
-        if n_right > 0
-        else 0
-    )
-
-    score = score_left + score_right
-    return score
-
-
-class SplitScoreMetrics(Enum):
-    # https://stackoverflow.com/questions/40338652/how-to-define-enum-values-that-are-functions
-    variance = partial(calc_variance)
-    entropy = partial(calc_entropy)
-    gini = partial(calc_gini_impurity)
-    # variance for split score because Friedman et al. 2001 in Algorithm 1
-    # step 4 minimize the squared error between actual and predicted dloss/dyhat
-    friedman_binary_classification = partial(calc_variance)
-    xgboost = partial(calc_xgboost_split_score)
-
-    def __call__(
-        self,
-        y: np.ndarray,
-        target_groups: np.ndarray,
-        yhat: np.ndarray = None,
-        g: np.ndarray = None,
-        h: np.ndarray = None,
-        growth_params: TreeGrowthParameters = None,
-    ) -> float:
-        return self.value(
-            y, target_groups, yhat=yhat, g=g, h=h, growth_params=growth_params
-        )
-
-
 @dataclass(config=ConfigDict(arbitrary_types_allowed=True))
 class BestSplit:
     score: StrictFloat
@@ -283,7 +95,7 @@ def find_best_split(
     yhat: np.ndarray = None,
     g: np.ndarray = None,
     h: np.ndarray = None,
-    growth_params: TreeGrowthParameters = None,
+    growth_params: utils.TreeGrowthParameters = None,
 ) -> BestSplit:
     """ """
     if len(np.unique(y)) == 1:
@@ -306,7 +118,7 @@ def find_best_split(
         if is_finite.all():
             for threshold in feature_values[1:]:
                 target_groups = feature_values < threshold
-                split_score = SplitScoreMetrics[measure_name](
+                split_score = scoring.SplitScoreMetrics[measure_name](
                     y,
                     target_groups,
                     yhat=yhat,
@@ -341,7 +153,7 @@ def find_best_split(
                             feature_values < threshold, is_finite
                         )
 
-                    split_score = SplitScoreMetrics[measure_name](
+                    split_score = scoring.SplitScoreMetrics[measure_name](
                         y,
                         target_groups,
                         yhat=yhat,
@@ -367,7 +179,9 @@ def find_best_split(
 
 
 def check_if_gain_insufficient(
-    best: BestSplit, parent_node: Node, growth_params: TreeGrowthParameters
+    best: BestSplit,
+    parent_node: Node,
+    growth_params: utils.TreeGrowthParameters,
 ) -> bool:
     if parent_node is None or parent_node.measure.value is None:
         return False, None
@@ -378,76 +192,6 @@ def check_if_gain_insufficient(
     return is_insufficient_gain, gain
 
 
-def leaf_weight_mean(
-    y: np.ndarray, growth_params: TreeGrowthParameters, **kwargs
-) -> float:
-    return np.mean(y)
-
-
-def leaf_weight_binary_classification_friedman2001(
-    y: np.ndarray, growth_params: TreeGrowthParameters, g: np.ndarray, **kwargs
-) -> float:
-    "Computes optimal leaf weight as in Friedman et al. 2001 Algorithm 5"
-
-    g_abs = np.abs(g)
-    gamma_jm = g.sum() / (g_abs * (2 - g_abs)).sum()
-    return gamma_jm
-
-
-def leaf_weight_xgboost(
-    y: np.ndarray,
-    growth_params: TreeGrowthParameters,
-    g: np.ndarray,
-    h: np.ndarray,
-    **kwargs,
-) -> float:
-    "Computes optimal leaf weight as in Chen et al. 2016 equation 5"
-
-    w = -g.sum() / (h + growth_params.lam).sum()
-    return w
-
-
-class LeafWeightSchemes(Enum):
-    # https://stackoverflow.com/questions/40338652/how-to-define-enum-values-that-are-functions
-    friedman_binary_classification = partial(
-        leaf_weight_binary_classification_friedman2001
-    )
-    variance = partial(leaf_weight_mean)
-    entropy = partial(leaf_weight_mean)
-    gini = partial(leaf_weight_mean)
-    xgboost = partial(leaf_weight_xgboost)
-
-    def __call__(
-        self,
-        y: np.ndarray,
-        growth_params: TreeGrowthParameters,
-        g: np.ndarray = None,
-        h: np.ndarray = None,
-    ) -> float:
-        return self.value(y, growth_params, g=g, h=h)
-
-
-def calc_leaf_weight(
-    y: np.ndarray,
-    measure_name: str,
-    growth_params: TreeGrowthParameters,
-    g: np.ndarray = None,
-    h: np.ndarray = None,
-) -> np.ndarray:
-    """Calculates the leaf weight, depending on the choice of measure_name.
-
-    This computation assumes all y values are part of the same leaf.
-    """
-
-    if len(y) == 0:
-        return None
-
-    weight_func = LeafWeightSchemes[measure_name]
-    leaf_weights = weight_func(y, growth_params, g=g, h=h)
-
-    return leaf_weights
-
-
 def grow_tree(
     X: np.ndarray,
     y: np.ndarray,
@@ -455,7 +199,7 @@ def grow_tree(
     yhat: np.ndarray = None,
     parent_node: Node = None,
     depth: int = 0,
-    growth_params: TreeGrowthParameters = None,
+    growth_params: utils.TreeGrowthParameters = None,
     g: np.ndarray = None,
     h: np.ndarray = None,
     **kwargs,
@@ -480,9 +224,11 @@ def grow_tree(
 
     leaf_weight, yhat, score = None, None, None
     if len(y) > 0:
-        leaf_weight = calc_leaf_weight(y, measure_name, growth_params, g=g, h=h)
+        leaf_weight = leafweights.calc_leaf_weight(
+            y, measure_name, growth_params, g=g, h=h
+        )
         yhat = leaf_weight * np.ones_like(y)
-        score = SplitScoreMetrics[measure_name](
+        score = scoring.SplitScoreMetrics[measure_name](
             y,
             np.ones_like(y, dtype=bool),
             yhat=yhat,
@@ -641,7 +387,7 @@ class DecisionTreeTemplate(base.BaseEstimator):
         self.lam = lam
 
     def _organize_growth_parameters(self):
-        self.growth_params_ = TreeGrowthParameters(**self.get_params())
+        self.growth_params_ = utils.TreeGrowthParameters(**self.get_params())
 
     def fit(
         self,
