+++ random_tree_models/utils.py
@@ -1,0 +1,18 @@
+from pydantic import (
+    ConfigDict,
+    Field,
+    StrictBool,
+    StrictFloat,
+    StrictInt,
+    StrictStr,
+)
+from pydantic.dataclasses import dataclass
+
+
+@dataclass
+class TreeGrowthParameters:
+    max_depth: StrictInt = 42
+    min_improvement: StrictFloat = 0.0
+    # xgboost lambda - multiplied with sum of squares of leaf weights
+    # see Chen et al. 2016 equation 2
+    lam: StrictFloat = 0.0
